requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.23
alembic==1.13.1
openai==1.51.0
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
redis==5.0.1
pandas==2.1.4
requests==2.31.0
stripe==7.8.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1